Crate linera_alloy_eips

linera-alloy-eips

Ethereum Improvement Proprosal (EIP) implementations.

Contains constants, helpers, and basic data structures for consensus EIPs.

Current support

EIP-1559

EIP-2718

EIP-2930

EIP-4788

EIP-4844

Re-exports

pub use eip1559::calc_next_block_base_fee;

pub use eip1898::BlockHashOrNumber;

pub use eip1898::BlockId;

pub use eip1898::BlockNumHash;

pub use eip1898::BlockNumberOrTag;

pub use eip1898::ForkBlock;

pub use eip1898::RpcBlockHash;

pub use eip4844::calc_blob_gasprice;

pub use eip4844::calc_excess_blob_gas;

Modules

eip1559

EIP-1559 constants, helpers, and types.

eip1898

eip2718

EIP-2718 traits.

eip2930

EIP-2930 types.

eip2935

Contains the history storage contract, first introduced in the Prague hardfork.

eip4788

EIP-4788 constants.

eip4844

EIP-4844 constants and helpers.

eip4895

EIP-4895 Withdrawal type and serde helpers.

eip6110

Contains Deposit types, first introduced in the Prague hardfork: https://github.com/ethereum/execution-apis/blob/main/src/engine/prague.md

eip7002

Contains the system contract and WithdrawalRequest types, first introduced in the Prague hardfork.

eip7685

EIP-7685: General purpose execution layer requests

merge

Constants related to the beacon chain consensus.