Module linera_alloy_eips::eip1559

EIP-1559 constants, helpers, and types.

Structs

BaseFeeParams

BaseFeeParams contains the config parameters that control block base fee computation

Constants

DEFAULT_BASE_FEE_MAX_CHANGE_DENOMINATOR

Base fee max change denominator as defined in EIP-1559

DEFAULT_ELASTICITY_MULTIPLIER

Elasticity multiplier as defined in EIP-1559

ETHEREUM_BLOCK_GAS_LIMIT

The default Ethereum block gas limit.

INITIAL_BASE_FEE

Initial base fee as defined in EIP-1559

MIN_PROTOCOL_BASE_FEE

The minimum tx fee below which the txpool will reject the transaction.

MIN_PROTOCOL_BASE_FEE_U256

Same as MIN_PROTOCOL_BASE_FEE but as a U256.

Functions

calc_next_block_base_fee

Calculate the base fee for the next block based on the EIP-1559 specification.