Module linera_alloy_eips::eip1898

Structs

BlockNumHash

Block number and hash.

HexStringMissingPrefixError

Thrown when a 0x-prefixed hex string was expected

ParseBlockHashOrNumberError

Error thrown when parsing a BlockHashOrNumber from a string.

RpcBlockHash

A block hash which may have a boolean requireCanonical field. If false, an RPC call should raise if a block matching the hash is not found. If true, an RPC call should additionally raise if the block is not in the canonical chain. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md#specification

Enums

BlockHashOrNumber

Either a block hash or a block number

BlockId

A Block Identifier. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

ParseBlockIdError

Error thrown when parsing a BlockId from a string.

ParseBlockNumberError

Error thrown when parsing a BlockNumberOrTag from a string.

Type Aliases

ForkBlock

Block number and hash of the forked block.