Trait linera_alloy_eips::eip2718::Decodable2718

pub trait Decodable2718: Sized {
    // Required methods
    fn typed_decode(ty: u8, buf: &mut &[u8]) -> Result<Self>;
    fn fallback_decode(buf: &mut &[u8]) -> Result<Self>;

    // Provided methods
    fn extract_type_byte(buf: &mut &[u8]) -> Option<u8> { ... }
    fn decode_2718(buf: &mut &[u8]) -> Result<Self> { ... }
    fn network_decode(buf: &mut &[u8]) -> Result<Self> { ... }
}

Decoding trait for EIP-2718 envelopes. These envelopes wrap a transaction or a receipt with a type flag.

Required Methods

fn typed_decode(ty: u8, buf: &mut &[u8]) -> Result<Self>

Decode the appropriate variant, based on the type flag.

This function is invoked by Self::decode_2718 with the type byte, and the tail of the buffer.

Note

This should be a simple match block that invokes an inner type’s RLP decoder.

fn fallback_decode(buf: &mut &[u8]) -> Result<Self>

Decode the default variant.

This function is invoked by Self::decode_2718 when no type byte can be extracted.

Provided Methods

fn extract_type_byte(buf: &mut &[u8]) -> Option<u8>

Extract the type byte from the buffer, if any. The type byte is the first byte, provided that that first byte is 0x7f or lower.

fn decode_2718(buf: &mut &[u8]) -> Result<Self>

Decode an EIP-2718 transaction into a concrete instance

fn network_decode(buf: &mut &[u8]) -> Result<Self>

Decode an EIP-2718 transaction in the network format.

The network format is the RLP encoded string consisting of the type-flag prepended to an opaque inner encoding. The inner encoding is RLP for all current Ethereum transaction types, but may not be in future versions of the protocol.

Object Safety

This trait is not object safe.