Module linera_alloy_eips::eip4844::builder

Builder and utils for the EIP-4844 Blob Transaction

Structs

PartialSidecar

A builder for creating a BlobTransactionSidecar.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

Traits

SidecarCoder

A strategy for coding and decoding data into sidecars. Coder instances are responsible for encoding and decoding data into and from the sidecar. They are called by the SidecarBuilder during the ingest, take, and (if c_kzg feature enabled) build methods.