Module linera_alloy_eips::eip7002

Contains the system contract and WithdrawalRequest types, first introduced in the Prague hardfork.

See also EIP-7002: Execution layer triggerable withdrawals

Structs

WithdrawalRequest

Represents an execution layer triggerable withdrawal request.

Constants

SYSTEM_ADDRESS

The caller to be used when calling the EIP-7002 withdrawal requests contract at the end of the block.

WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS

The address for the EIP-7002 withdrawal requests contract.

WITHDRAWAL_REQUEST_TYPE

The EIP-7685 request type for withdrawal requests.

Statics

WITHDRAWAL_REQUEST_PREDEPLOY_CODE

The code for the EIP-7002 withdrawal requests contract.